Move to AC & Clean up
