+++ Assets/ActivateCode/CH/Scripts/Monster/Abstract.meta
