+++ ProjectSettings/TagManager.asset
@@ -6,7 +6,6 @@ TagManager:
   tags:
   - CharacterHand
   - Character
-  - PlayerAttackHitbox
   layers:
   - Default
   - TransparentFX
